Huh, workaround bug in virtual dispatch to write_value().
FossilOrigin-Name: 66f51f3d3841878fbac99640d57d612fc9ebd30a376f6776ff5daa85766e9a93

+++ components/SampleSummaryComponent/src/FlatFileOutputter.cpp
@@ -120,7 +120,7 @@ namespace sample_stats {
 		for( std::size_t i = 0; i < m_sample_indices.size(); ++i ) {
 			std::size_t const sample_index = m_sample_indices[ i ] ;
 			genfile::VariantEntry const& sample = m_samples[ sample_index ] ;
-			(*m_sink) << sample << sample_index ;
+			(*m_sink) << sample << uint64_t( sample_index ) ;
 			VariableMap::right_const_iterator
 				var_i = m_variables.right.begin(),
 				end_var_i = m_variables.right.end() ;
